# BlueBank/src/main/java/com/squad/dezktop/model/ContaModel.java
package com.squad.dezktop.model;

import java.math.BigDecimal;

import javax.persistence.Entity;
import javax.persistence.Id;
import javax.persistence.JoinColumn;
import javax.persistence.OneToOne;
import javax.persistence.Table;
import javax.validation.constraints.Size;

import com.fasterxml.jackson.annotation.JsonIgnore;

@Entity
@Table(name = "conta")
public class ContaModel 
{
	@Id
	@Size(min = 10, max = 10)
	private String numeroConta;

	@Size(min = 4, max = 6)
	private String agencia;
	
	private BigDecimal saldo;

	@JsonIgnore
	@OneToOne
	@JoinColumn(name = "cliente_cpf", referencedColumnName = "cpf")
	private ClienteModel cliente;
	
	public String getNumeroConta() {
		return numeroConta;
	}
	public void setNumeroConta(String numeroConta) {
		this.numeroConta = numeroConta;
	}
	public String getAgencia() {
		return agencia;
	}
	public void setAgencia(String agencia) {
		this.agencia = agencia;
	}
	public BigDecimal getSaldo() {
		return saldo;
	}
	public void setSaldo(BigDecimal saldo) {
		this.saldo = saldo;
	}
	public ClienteModel getCliente() {
		return cliente;
	}
	public void setCliente(ClienteModel cliente) {
		this.cliente = cliente;
	}
	public String getClienteCPF() {
		return cliente.getCpf();
	}
	
	
}