# swagger/API_Bluebank_Clientes_Contas.json
{
    "swagger": "2.0",
    "info": {
        "description": "O projeto Bluebank tem como objetivo criar uma API para gerenciar as transações de um banco fictício utilizando Java, MySQL e AWS.\nNesta página, você encontra as funções relacionadas ao nosso primeiro microsserviço: Clientes e Contas.\nAqui, você pode cadastrar e deletar clientes e contas, atualizar dados dos clientes e listar clientes e contas.\nVocê pode achar mais informações sobre o processo de desenvolvimento, no nosso repositório no GitHub no link abaixo :)",
        "version": "1.0.0",
        "title": "API Bluebank: Microsserviço de clientes e contas",
        "contact": {
            "name": "Bluebank repository",
            "url": "https://github.com/ivancmc/bluebank"
        },
        "license": {
            "name": "Apache License Version 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0"
        }
    },
    "host": "ec2-3-82-216-0.compute-1.amazonaws.com:8101",
    "basePath": "/",
    "tags": [
        {
            "name": "Cliente",
            "description": " "
        },
        {
            "name": "Conta",
            "description": " "
        }
    ],
    "paths": {
        "/cliente": {
            "post": {
                "tags": [
                    "Cliente"
                ],
                "summary": "Cadastra um novo cliente.",
                "description": "Para cadastrar um novo cliente é necessário inserir algumas informações pessoais como: \n- Nome completo, endereço, número de telefone, data de nascimento, CPF, RG, email e senha.\n- Os dados bancários são gerados automaticamente.",
                "operationId": "postUsingPOST",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "cliente",
                        "description": "cliente",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ClienteModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "201": {
                        "description": "Criado"
                    },
                    "400": {
                        "description": "Erro inesperado"
                    },
                    "401": {
                        "description": "Não autorizado"
                    }
                },
                "deprecated": false
            }
        },
        "/cliente/{cpf}": {
            "get": {
                "tags": [
                    "Cliente"
                ],
                "summary": "Retorna um cliente.",
                "operationId": "getByCpfUsingGET",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "cpf",
                        "in": "path",
                        "description": "cpf",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ClienteModel"
                        }
                    },
                    "404": {
                        "description": "Não encontrado"
                    }
                },
                "deprecated": false
            },
            "put": {
                "tags": [
                    "Cliente"
                ],
                "summary": "Atualiza os dados de um cliente.",
                "description": "Para atualizar os dados do cliente é necessário inserir o número de CPF no campo indicado.\n\n- Não é possível alterar o CPF ou os dados bancários.",
                "operationId": "updateUsingPUT",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "cliente",
                        "description": "cliente",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ClienteModel"
                        }
                    },
                    {
                        "name": "cpf",
                        "in": "path",
                        "description": "cpf",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ClienteModel"
                        }
                    },
                    "400": {
                        "description": "Erro inesperado"
                    },
                    "404": {
                        "description": "Não encontrado"
                    }
                },
                "deprecated": false
            },
            "delete": {
                "tags": [
                    "Cliente"
                ],
                "summary": "Deleta um cliente.",
                "operationId": "deleteUsingDELETE",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "cpf",
                        "in": "path",
                        "description": "cpf",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Erro inesperado"
                    },
                    "404": {
                        "description": "Não encontrado"
                    }
                },
                "deprecated": false
            }
        },
        "/clientes": {
            "get": {
                "tags": [
                    "Cliente"
                ],
                "summary": "Retorna uma lista com todos os clientes.",
                "operationId": "getAllUsingGET",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ClienteModel"
                            }
                        }
                    },
                    "404": {
                        "description": "Não encontrado"
                    }
                },
                "deprecated": false
            }
        },
        "/conta/cpf/{cpf}": {
            "get": {
                "tags": [
                    "Conta"
                ],
                "summary": "Retorna uma conta ao buscar por CPF.",
                "operationId": "getByCpfUsingGET_1",
                "produces": [
                    "*/*",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "cpf",
                        "in": "path",
                        "description": "cpf",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ContaModel"
                        }
                    },
                    "404": {
                        "description": "Não encontrado"
                    }
                },
                "deprecated": false
            }
        },
        "/conta/numero/{numero}": {
            "get": {
                "tags": [
                    "Conta"
                ],
                "summary": "Retorna uma conta ao buscar por número da conta.",
                "operationId": "getByNumeroUsingGET",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "numero",
                        "in": "path",
                        "description": "numero",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ContaModel"
                        }
                    },
                    "404": {
                        "description": "Não encontrado"
                    }
                },
                "deprecated": false
            }
        },
        "/conta/{numero}/credita/{valor}": {
            "put": {
                "tags": [
                    "Conta"
                ],
                "summary": "Adiciona valor na conta especificada. Só deve ser chamado pelo serviço de transações e não deve ser requisitado diretamente.",
                "operationId": "creditaUsingPUT",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "numero",
                        "in": "path",
                        "description": "numero",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "valor",
                        "in": "path",
                        "description": "valor",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ContaModel"
                        }
                    },
                    "400": {
                        "description": "Erro inesperado"
                    },
                    "404": {
                        "description": "Não encontrado"
                    }
                },
                "deprecated": false
            }
        },
        "/conta/{numero}/debita/{valor}": {
            "put": {
                "tags": [
                    "Conta"
                ],
                "summary": "Diminui valor na conta especificada. Só deve ser chamado pelo serviço de transações e não deve ser requisitado diretamente.",
                "operationId": "debitaUsingPUT",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "numero",
                        "in": "path",
                        "description": "numero",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "valor",
                        "in": "path",
                        "description": "valor",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ContaModel"
                        }
                    },
                    "400": {
                        "description": "Erro inesperado"
                    },
                    "404": {
                        "description": "Não encontrado"
                    }
                },
                "deprecated": false
            }
        }
    },
    "definitions": {
        "ClienteModel": {
            "type": "object",
            "properties": {
                "conta": {
                    "$ref": "#/definitions/ContaModel"
                },
                "cpf": {
                    "type": "string"
                },
                "dataNascimento": {
                    "type": "string",
                    "format": "date-time"
                },
                "email": {
                    "type": "string"
                },
                "endereco": {
                    "type": "string"
                },
                "nome": {
                    "type": "string"
                },
                "rg": {
                    "type": "string"
                },
                "senha": {
                    "type": "string"
                },
                "telefone": {
                    "type": "string"
                }
            },
            "title": "ClienteModel"
        },
        "ContaModel": {
            "type": "object",
            "properties": {
                "agencia": {
                    "type": "string"
                },
                "clienteCPF": {
                    "type": "string"
                },
                "numeroConta": {
                    "type": "string"
                },
                "saldo": {
                    "type": "number"
                }
            },
            "title": "ContaModel"
        }
    }
}